fix: split basic catalog unit tests into per-component files
Split simple-components.spec.ts and complex-components.spec.ts into
individual test files, one per component. This produces smaller files
with self-contained test setup, reducing complexity and merge conflict
likelihood.

New files created:
- divider.component.spec.ts
- image.component.spec.ts
- icon.component.spec.ts
- video.component.spec.ts
- audio-player.component.spec.ts
- card.component.spec.ts
- check-box.component.spec.ts
- choice-picker.component.spec.ts
- slider.component.spec.ts
- date-time-input.component.spec.ts
- list.component.spec.ts
- tabs.component.spec.ts
- modal.component.spec.ts

Deleted files:
- simple-components.spec.ts
- complex-components.spec.ts

Fixes #1403
