+++ renderers/angular/src/v0_9/catalog/basic/divider.component.spec.ts
@@ -1,0 +1,83 @@
+/**
+ * Copyright 2026 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {ComponentFixture, TestBed} from '@angular/core/testing';
+import {DividerComponent} from './divider.component';
+import {A2uiRendererService} from '../../core/a2ui-renderer.service';
+import {ComponentBinder} from '../../core/component-binder.service';
+import {setComponentProps, createBoundProperty, ComponentToProps} from '../../core/test-utils';
+
+describe('DividerComponent', () => {
+  let component: DividerComponent;
+  let fixture: ComponentFixture<DividerComponent>;
+  let defaultProps: ComponentToProps<DividerComponent>;
+
+  beforeEach(async () => {
+    const mockRendererService = {
+      surfaceGroup: {
+        getSurface: jasmine.createSpy('getSurface').and.returnValue({
+          componentsModel: new Map(),
+          catalog: {
+            id: 'mock-catalog',
+            components: new Map(),
+          },
+        }),
+      },
+    };
+    const mockBinder = jasmine.createSpyObj('ComponentBinder', ['bind']);
+
+    await TestBed.configureTestingModule({
+      imports: [DividerComponent],
+      providers: [
+        {provide: A2uiRendererService, useValue: mockRendererService},
+        {provide: ComponentBinder, useValue: mockBinder},
+      ],
+    }).compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(DividerComponent);
+    component = fixture.componentInstance;
+    fixture.componentRef.setInput('surfaceId', 'test-surface');
+    fixture.componentRef.setInput('dataContextPath', '/');
+
+    defaultProps = {
+      axis: createBoundProperty('horizontal' as const),
+    };
+    setComponentProps(fixture, defaultProps);
+  });
+
+  it('should create', () => {
+    fixture.detectChanges();
+    expect(component).toBeTruthy();
+  });
+
+  it('should render horizontal divider by default', () => {
+    fixture.detectChanges();
+    const divider = fixture.nativeElement.querySelector('.a2ui-divider');
+    expect(divider.classList).toContain('horizontal');
+  });
+
+  it('should render vertical divider', () => {
+    setComponentProps(fixture, {
+      ...defaultProps,
+      axis: createBoundProperty('vertical' as const),
+    });
+    fixture.detectChanges();
+    const divider = fixture.nativeElement.querySelector('.a2ui-divider');
+    expect(divider.classList).toContain('vertical');
+  });
+});
